Clean up a few minor things
